pythonBackendAPI/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
PyJWT==2.6.0
psycopg2-binary==2.9.5

